refactror: 機能を分割
